- added robins patch for #65 - build system supports now the build of kismac images (./compile.command image)

+++ KisMACng/compile.command
@@ -1,11 +1,48 @@
 #!/bin/bash
 
-configuration=Deployment
+if [ "x$1" == "x--help" ]; then
+  echo "Usage ./compile.command [Universal|Development|Deployment] [image]"
+  exit 0
+fi
 
-echo "Checking for required enviroment..."
+case $1 in
+  "Universal")
+    configuration=Universal;;
+  "Development")
+    configuration=Development;;
+  "Deployment" | "" | "image")
+    configuration=Deployment;;
+  *)
+    echo "Usage ./compile.command [Universal|Development|Deployment] [image]"
+    exit 1
+esac
+
+echo "Building for configuration... $configuration"
+
+echo -n "Building install image... "
+if [ "x$1" == "ximage" ] || [ "x$2" == "ximage" ]; then
+  echo "YES"
+  BUILD_IMAGE=1
+else
+  echo "NO"
+	BUILD_IMAGE=0
+fi
+
+echo -n "Checking for required enviroment... "
 if ! [ -x /usr/bin/tar ]; then
 	echo "/usr/bin/tar not found! Make sure you installed the BSD subsystem!"
 	exit 1
+fi
+
+if ! [ -x /usr/bin/xcodebuild ]; then
+	echo "Could not find a valid XCode version!"
+	exit 1
+fi
+
+XCODEVERSION=`xcodebuild -version  | sed s/.*DevToolsCore-// | sed  s/..\;.*//`
+if [ $XCODEVERSION \< 658 ]; then
+  echo "XCode Version is too old!"
+  exit 1
 fi
 
 if echo $0 | grep " " > /dev/null; then
@@ -13,26 +50,38 @@ if echo $0 | grep " " > /dev/null; then
 	exit 1
 fi
 
-cd "`dirname "$0"`"
+echo "ok"
 
-echo "Decompressing UnitTest bundle..."
+echo -n "Decompressing UnitTest bundle... "
 mkdir "./build/KisMACUnitTest.bundle/Contents/Frameworks" 2>/dev/null
 cd UnitTest
 ln -s "../build/KisMACUnitTest.bundle/Contents/Frameworks" . 2>/dev/null
 tar -xjf UnitKit.tbz 2>/dev/null
+if ! [ -d UnitKit.framework ]; then
+  echo "*FAILED*"
+  exit 1
+fi
+echo "ok"
 cd ..
 
-echo "Decompressing Growl framework..."
+echo -n "Decompressing Growl framework... "
 cd Resources
 tar -xzf growl.tgz
+if ! [ -d Growl.framework ]; then
+  echo "*FAILED*"
+  exit 1
+fi
+echo "ok"
+
 cd ..
 
-echo "Determine Subversion Revision..."
-val=`svnversion -n .`
-sed -e "s/\\\$Revision.*\\\$/\\\$Revision: $val\\\$/" Resources/Info.plist.templ > Resources/Info.plist
-sed -e "s/\\\$Revision.*\\\$/\\\$Revision: $val\\\$/" Resources/Strings/English.lproj/InfoPlist.strings.templ > Resources/Strings/English.lproj/InfoPlist.strings
+echo -n "Determine Subversion Revision... "
+SVNVERS=`svnversion -n .`
+echo $SVNVERS
+sed -e "s/\\\$Revision.*\\\$/\\\$Revision: $SVNVERS\\\$/" Resources/Info.plist.templ > Resources/Info.plist
+sed -e "s/\\\$Revision.*\\\$/\\\$Revision: $SVNVERS\\\$/" Resources/Strings/English.lproj/InfoPlist.strings.templ > Resources/Strings/English.lproj/InfoPlist.strings
 
-echo "Preparing Enviroment..."
+echo -n "Preparing Enviroment... "
 if [ -f compile.log ]; then
   rm compile.log
 fi
@@ -42,6 +91,8 @@ mkdir "./Subprojects/files" 2>/dev/null
 cd "./Subprojects/files"
 rm -rf *.framework 2>/dev/null
 cd ..
+
+echo "ok"
 
 echo -n "Building MACJack driver... "
 cd MACJack
@@ -112,6 +163,57 @@ cd ../..
 if ! xcodebuild -target KisMAC -configuration $configuration >> compile.log; then
         exit 1
 else
-        echo "ok"
+         echo "ok"
 fi
 
+if [ $BUILD_IMAGE == 1 ]; then
+  echo -n "Generating KisMAC Disk Image..."
+  
+  if [ -f /Volumes/KisMAC ]; then
+    	if hdiutil unmount /Volumes/KisMAC 2>/dev/null; then
+        echo "*FAILED* Could not unmount loaded KisMAC Volume."
+        exit 1
+      fi
+      sleep 10
+  fi
+  
+	cp image/KisMACraw.dmg image/KisMAC.dmg
+	hdiutil attach image/KisMAC.dmg > /dev/null
+	
+	sleep 1
+	
+	if ! [ -d /Volumes/KisMAC ]; then
+    echo " *FAILED* Could not mount KisMAC Volume."
+    exit 1
+  fi 
+  
+	cp -r "Subprojects/KisMACInstaller/build/$configuration/KisMAC Installer.app/Contents" "/Volumes/KisMAC/KisMAC Installer.app"
+  cp "image/WirelessDriver" "/Volumes/KisMAC/KisMAC Installer.app/Contents/Resources"
+  
+  cp -r build/$configuration/KisMAC.app image
+  cd "image/KisMAC.app"
+	rm `find . -type f -name .DS_Store` 2>/dev/null
+	rm -rf `find . -name .svn`
+	tar -czf "/Volumes/KisMAC/KisMAC Installer.app/Contents/Resources/KisMAC.tgz" *>/dev/null
+	cd ..
+	rm -rf KisMAC.app
+	cd ..
+  	
+	rm `find "/Volumes/KisMAC/KisMAC Installer.app" -type f -name .DS_Store` 2>/dev/null
+	rm -rf `find "/Volumes/KisMAC/KisMAC Installer.app" -name .svn`
+
+  while [ -d  /Volumes/KisMAC ]; do
+    hdiutil detach /Volumes/KisMAC >/dev/null
+    sleep 1
+    echo -n .
+  done
+  
+	#trim dumb characters from SVN version
+	SVNVERS=`echo $SVNVERS | sed 's/.*://g'`
+  SVNVERS=`echo $SVNVERS | sed 's/\]//g'`
+  
+	zip -9 KisMACR$SVNVERS.zip image/KisMAC.dmg >/dev/null
+	
+	rm image/KisMAC.dmg
+	echo "ok"
+fi
